# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Polymarket Broker API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/webhooks": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Create Webhook",
        "description": "Register a new webhook endpoint.",
        "operationId": "create_webhook_api_v1_webhooks_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "webhooks"
        ],
        "summary": "List Webhooks",
        "description": "List all webhooks for the authenticated user.",
        "operationId": "list_webhooks_api_v1_webhooks_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebhookResponse"
                  },
                  "title": "Response List Webhooks Api V1 Webhooks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/developer/usage": {
      "get": {
        "tags": [
          "developer"
        ],
        "summary": "Get Usage",
        "description": "API usage stats for today.",
        "operationId": "get_usage_api_v1_developer_usage_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/developer/billing": {
      "get": {
        "tags": [
          "developer"
        ],
        "summary": "Get Billing",
        "description": "Current billing status.",
        "operationId": "get_billing_api_v1_developer_billing_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/developer/billing/upgrade": {
      "post": {
        "tags": [
          "developer"
        ],
        "summary": "Upgrade Tier",
        "description": "Upgrade subscription tier.",
        "operationId": "upgrade_tier_api_v1_developer_billing_upgrade_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpgradeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpgradeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/developer/webhooks": {
      "get": {
        "tags": [
          "developer"
        ],
        "summary": "Get Webhook Health",
        "description": "Webhook health status for all registered webhooks.",
        "operationId": "get_webhook_health_api_v1_developer_webhooks_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebhookHealthResponse"
                  },
                  "title": "Response Get Webhook Health Api V1 Developer Webhooks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/developer/risk/reset": {
      "post": {
        "tags": [
          "developer"
        ],
        "summary": "Reset Risk",
        "description": "Reset daily drawdown circuit breaker.",
        "operationId": "reset_risk_api_v1_developer_risk_reset_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskResetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BillingResponse": {
        "properties": {
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "next_billing_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Billing Date"
          },
          "amount_due_cents": {
            "type": "integer",
            "title": "Amount Due Cents"
          }
        },
        "type": "object",
        "required": [
          "tier",
          "next_billing_date",
          "amount_due_cents"
        ],
        "title": "BillingResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskResetResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "circuit_breaker_reset": {
            "type": "boolean",
            "title": "Circuit Breaker Reset"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "circuit_breaker_reset",
          "message"
        ],
        "title": "RiskResetResponse"
      },
...
      "UpgradeRequest": {
        "properties": {
          "tier": {
            "type": "string",
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "tier"
        ],
        "title": "UpgradeRequest"
      },
      "UpgradeResponse": {
        "properties": {
          "previous_tier": {
            "type": "string",
            "title": "Previous Tier"
          },
          "new_tier": {
            "type": "string",
            "title": "New Tier"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "previous_tier",
          "new_tier",
          "message"
        ],
        "title": "UpgradeResponse"
      },
      "UsageResponse": {
        "properties": {
          "calls_today": {
            "type": "integer",
            "title": "Calls Today"
          },
          "calls_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calls Remaining"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "reset_at": {
            "type": "string",
            "format": "date-time",
            "title": "Reset At"
          }
        },
        "type": "object",
        "required": [
          "calls_today",
          "calls_remaining",
          "tier",
          "reset_at"
        ],
        "title": "UsageResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebhookCreateRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          },
          "secret": {
            "type": "string",
            "title": "Secret"
          }
        },
        "type": "object",
        "required": [
          "url",
          "events",
          "secret"
        ],
        "title": "WebhookCreateRequest"
      },
      "WebhookHealthResponse": {
        "properties": {
          "webhook_id": {
            "type": "string",
            "title": "Webhook Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "failure_count": {
            "type": "integer",
            "title": "Failure Count"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "webhook_id",
          "url",
          "status",
          "failure_count",
          "events"
        ],
        "title": "WebhookHealthResponse"
      },
      "WebhookResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "failure_count": {
            "type": "integer",
            "title": "Failure Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "events",
          "status",
          "failure_count",
          "created_at"
        ],
        "title": "WebhookResponse"
      }
    }
  }
}